--- a/be/src/vec/sink/writer/vtablet_writer.h
+++ b/be/src/vec/sink/writer/vtablet_writer.h
@@ -553,10 +553,13 @@ private:
     using ChannelDistributionPayload = std::vector<std::unordered_map<VNodeChannel*, Payload>>;
 
     Status _init(RuntimeState* state, RuntimeProfile* profile);
-    // payload for each row
-    void _generate_row_distribution_payload(ChannelDistributionPayload& payload,
-                                            const VOlapTablePartition* partition,
-                                            uint32_t tablet_index, int row_idx, size_t row_cnt);
+
+    // payload for every row
+    void _generate_row_distribution_payload(ChannelDistributionPayload& channel_to_payload,
+                                            const std::vector<VOlapTablePartition*>& partitions,
+                                            const std::vector<uint32_t>& tablet_indexes,
+                                            const std::vector<bool>& skip, size_t row_cnt);
+
     Status _single_partition_generate(RuntimeState* state, vectorized::Block* block,
                                       ChannelDistributionPayload& channel_to_payload,
                                       size_t num_rows, bool has_filtered_rows);